requirements.txt
agentscope>=1.0.0
dashscope>=1.14.0
chromadb>=0.4.0
pypdf2>=3.0.0
python-docx>=0.8.11
markdown>=3.5.0
pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
pyyaml>=6.0
chardet>=4.0.0